Add Estate and select Estate pages
