+++ lib/routes/routes.dart
@@ -1,24 +1,25 @@
 // import 'pages/splash_screen.dart';
 
 import 'package:flutter/material.dart';
+import 'package:gateapp/pages/Add_Estate.dart';
 import 'package:gateapp/pages/Pager/pager.dart';
-import 'package:gateapp/pages/manage_address.dart';
-import 'package:gateapp/pages/splash_screen.dart';
+import 'package:gateapp/pages/Select_Estate.dart';
+
 
 class Routes {
   static Route<dynamic> generateRoute(RouteSettings settings) {
     switch (settings.name) {
       case '/':
-        return MaterialPageRoute(builder: (context) => SplashScreen());
+        return MaterialPageRoute(builder: (context) => ManageAddress());
 
-      case '/splash':
-        return MaterialPageRoute(builder: (context) => SplashScreen());
-
+     
       case '/pager':
         return MaterialPageRoute(builder: (context) => Pager());
 
       case '/manage-address':
         return MaterialPageRoute(builder: (context) => ManageAddress());
+         case '/add-estate':
+        return MaterialPageRoute(builder: (context) => AddEstate());
 
       default:
         return MaterialPageRoute(
